[FIX] Type Issues

+++ src/UIHelper/Scale/ScaleFormGUI.php
@@ -114,7 +114,7 @@ class ScaleFormGUI extends ilPropertyFormGUI
             $old = $this->getArrayFromPostComplex(self::FIELD_NAME . '_old');
             if (is_array($old['value'])) {
                 foreach ($old['value'] as $k => $v) {
-                    $obj = new ScaleUnit($this->db, str_replace('id_', '', (string) $k));
+                    $obj = new ScaleUnit($this->db, (int) str_replace('id_', '', (string) $k));
                     if ($v !== false && $v !== null && $v !== '') {
                         $obj->setTitle($old['title'][$k]);
                         $obj->setValue((int) $v);
